Upgrade add-layers sub-modules to ng9
