+++ components/add-layers/add-layers.module.ts
@@ -1,7 +1,6 @@
-import {CUSTOM_ELEMENTS_SCHEMA} from '@angular/core';
+import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
-import {NgModule} from '@angular/core';
 import {TranslateModule} from '@ngx-translate/core';
 
 import {HsAddLayersComponent} from './add-layers.component';
