+++ components/add-layers/vector/add-layers-vector.module.ts
@@ -1,43 +1,15 @@
-import '../../language';
-import * as angular from 'angular';
+import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {FormsModule} from '@angular/forms';
+import {TranslateModule} from '@ngx-translate/core';
+
 import {HsAddLayersVectorComponent} from './add-layers-vector.component';
-import {HsAddLayersVectorService} from './add-layers-vector.service';
-import {HsVectorUrlParserService} from './add-layers-vector-url-parser.service';
 
-/**
- * @param service
- * @param layoutService
- * @namespace hs.addLayersVector
- * @memberof hs
- */
-angular
-  .module('hs.addLayersVector', [
-    'hs.styles',
-    'hs.utils',
-    'hs.language',
-    'hs.layout',
-    'hs.map'
-  ])
-  /**
-   * @memberof hs.addLayersVector
-   * @ngdoc directive
-   * @name hs.addLayersVector
-   * @description TODO
-   */
-  .component('hs.addLayersVector', HsAddLayersVectorComponent)
-
-  /**
-   * @memberof hs.addLayersVector
-   * @ngdoc service
-   * @name HsAddLayersVectorService
-   * @description Service handling adding nonwms OWS services or files. Handles also drag and drop addition.
-   */
-  .service('HsAddLayersVectorService', HsAddLayersVectorService)
-
-  /**
-   * @memberof hs.addLayersVector
-   * @ngdoc service
-   * @name hs.ddLayersVectorUrlParser.service
-   * @description Service handling loading of vector layers through url params
-   */
-  .factory('hs.addLayersVectorUrlParser.service', HsVectorUrlParserService);
+@NgModule({
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+  imports: [CommonModule, FormsModule, TranslateModule],
+  exports: [],
+  declarations: [/*HsAddLayersVectorComponent*/],
+  providers: [],
+})
+export class HsAddLayersVectorModule {}
