+++ components/add-layers/wfs/add-layers-wfs.module.ts
@@ -1,31 +1,15 @@
-import '../../../common/get-capabilities.module';
-import '../../utils';
-import * as angular from 'angular';
+import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {FormsModule} from '@angular/forms';
+import {TranslateModule} from '@ngx-translate/core';
+
 import {HsAddLayersWfsComponent} from './add-layers-wfs.component';
-import {HsGetCapabilitiesErrorComponent} from '../capabilities-error.component';
 
-/**
- * @namespace hs.addLayersWfs
- * @memberOf hs
- */
-angular
-  .module('hs.addLayersWfs', ['hs.utils', 'hs.getCapabilities'])
-
-  /**
-   * @name hs.addLayersWfs.capabalitiesErrorDirective
-   * @ngdoc directive
-   * @memberOf hs.addLayersWfs
-   * @description Display GetCapabilities error dialog template
-   */
-  .component(
-    'hs.addLayersWfs.capabilitiesErrorDirective',
-    HsGetCapabilitiesErrorComponent
-  )
-
-  /**
-   * @name hs.addLayersWfs
-   * @ngdoc controller
-   * @memberOf hs.addLayersWfs
-   * @description Controller for displaying and setting parameters for Wfs and its layers, which will be added to map afterwards
-   */
-  .component('hs.addLayersWfs', HsAddLayersWfsComponent);
+@NgModule({
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+  imports: [CommonModule, FormsModule, TranslateModule],
+  exports: [],
+  declarations: [/*HsAddLayersWfsComponent*/],
+  providers: [],
+})
+export class HsAddLayersWfsModule {}
